Organise UI directory
