+++ src/main/java/seedu/address/ui/home/BrowserPanel.java
@@ -1,4 +1,4 @@
-package seedu.address.ui;
+package seedu.address.ui.home;
 
 import java.net.URL;
 import java.util.logging.Logger;
@@ -16,6 +16,7 @@ import seedu.address.commons.events.ui.CardPanelSelectionChangedEvent;
 import seedu.address.commons.events.ui.PersonPanelSelectionChangedEvent;
 import seedu.address.model.card.Card;
 import seedu.address.model.person.Person;
+import seedu.address.ui.UiPart;
 
 /**
  * The Browser Panel of the App.
@@ -26,7 +27,7 @@ public class BrowserPanel extends UiPart<Region> {
     public static final String SEARCH_PAGE_URL =
             "https://www.google.com.sg/search?q=";
 
-    private static final String FXML = "BrowserPanel.fxml";
+    private static final String FXML = "/home/BrowserPanel.fxml";
 
     private final Logger logger = LogsCenter.getLogger(getClass());
