+++ src/main/java/seedu/address/ui/home/CardView.java
@@ -1,4 +1,4 @@
-package seedu.address.ui;
+package seedu.address.ui.home;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
@@ -6,13 +6,14 @@ import javafx.scene.layout.FlowPane;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Region;
 import seedu.address.model.card.Card;
+import seedu.address.ui.UiPart;
 
 /**
  * An UI component that displays information of a {@code Card}.
  */
 public class CardView extends UiPart<Region> {
 
-    private static final String FXML = "CardView.fxml";
+    private static final String FXML = "/home/CardView.fxml";
 
     /**
      * Note: Certain keywords such as "location" and "resources" are reserved keywords in JavaFX.
