+++ src/main/resources/view/home/BrowserPanel.fxml
