+++ src/main/resources/view/home/CardView.fxml
