+++ src/test/java/systemtests/HelpCommandSystemTest.java
@@ -14,8 +14,8 @@ import guitests.guihandles.HelpWindowHandle;
 import seedu.address.logic.commands.DeleteCommand;
 import seedu.address.logic.commands.HelpCommand;
 import seedu.address.logic.commands.SelectCommand;
-import seedu.address.ui.BrowserPanel;
 import seedu.address.ui.StatusBarFooter;
+import seedu.address.ui.home.BrowserPanel;
 
 /**
  * A system test class for the help window, which contains interaction with other UI components.
